+++ lib/napa.rb
@@ -17,7 +17,7 @@ require 'napa/stats'
 require 'napa/active_record_extensions/filter_by_hash'
 require 'napa/grape_extensions/error_formatter'
 require 'napa/grape_extensions/error_presenter'
-require 'napa/exception_catchers'
+require 'napa/grape_extenders'
 require 'napa/middleware/logger'
 require 'napa/middleware/app_monitor'
 require 'napa/middleware/authentication'
